requirements.txt
PodSixNet
PySimpleGUI
PYYaml